Move mixin classes
