+++ src/clauses/mixins/sub-clauses/WithSet.ts
@@ -19,9 +19,9 @@
 
 import type { SetParam } from "../../sub-clauses/Set";
 import { SetClause } from "../../sub-clauses/Set";
-import { ClauseMixin } from "../ClauseMixin";
+import { Mixin } from "../Mixin";
 
-export abstract class WithSet extends ClauseMixin {
+export abstract class WithSet extends Mixin {
     protected setSubClause: SetClause | undefined;
 
     /** Append a `SET` clause. Allowing to assign variable properties to values.
